migrate Page method here
