+++ OneMore/Models/Page.cs
@@ -240,29 +240,6 @@ namespace River.OneMoreAddIn.Models
 
 
 		/// <summary>
-		/// Adds the given content after the selected insertion point; this will not
-		/// replace selected regions.
-		/// </summary>
-		/// <param name="content">The content to add</param>
-		public void AddNextParagraph(XElement content)
-		{
-			InsertParagraph(content, false);
-		}
-
-
-		public void AddNextParagraph(params XElement[] content)
-		{
-			// consumer will build content array in document-order but InsertParagraph inserts
-			// just prior to the insertion point which will reverse the order of content items
-			// so insert them in reverse order intentionally so they show up correctly
-			for (var i = content.Length - 1; i >= 0; i--)
-			{
-				InsertParagraph(content[i], false);
-			}
-		}
-
-
-		/// <summary>
 		/// Adds the given QuickStyleDef element in the proper document order, just after
 		/// the TagDef elements if there are any
 		/// </summary>
@@ -1145,43 +1122,6 @@ namespace River.OneMoreAddIn.Models
 			}
 
 			return false;
-		}
-
-
-		/// <summary>
-		/// Adds the given content immediately before or after the selected insertion point;
-		/// this will not replace selected regions.
-		/// </summary>
-		/// <param name="content">The content to insert</param>
-		/// <param name="before">
-		/// If true then insert before the insertion point; otherwise insert after the insertion point
-		/// </param>
-		public void InsertParagraph(XElement content, bool before = true)
-		{
-			var current = Root.Descendants(Namespace + "OE").LastOrDefault(e =>
-				e.Elements(Namespace + "T").Attributes("selected").Any(a => a.Value == "all"));
-
-			if (current != null)
-			{
-				if (content.Name.LocalName != "OE")
-				{
-					content = new XElement(Namespace + "OE", content);
-				}
-
-				if (before)
-					current.AddBeforeSelf(content);
-				else
-					current.AddAfterSelf(content);
-			}
-		}
-
-
-		public void InsertParagraph(params XElement[] content)
-		{
-			foreach (var e in content)
-			{
-				InsertParagraph(e, false);
-			}
 		}
